Implement MVVM, search item

+++ app/src/main/java/com/grjt/mercadolibre/data/model/ReferencePrices.kt
@@ -8,7 +8,7 @@ data class ReferencePrices (
   @SerializedName("id"                    ) var id                  : String?           = null,
   @SerializedName("type"                  ) var type                : String?           = null,
   @SerializedName("conditions"            ) var conditions          : Conditions?       = Conditions(),
-  @SerializedName("amount"                ) var amount              : Int?              = null,
+  @SerializedName("amount"                ) var amount              : Double?              = null,
   @SerializedName("currency_id"           ) var currencyId          : String?           = null,
   @SerializedName("exchange_rate_context" ) var exchangeRateContext : String?           = null,
   @SerializedName("tags"                  ) var tags                : ArrayList<String> = arrayListOf(),
